rename segment->segment_id and segemnt_ops->segment
