+++ src/metadata/metadata_raw_volatile.c
@@ -4,7 +4,7 @@
  */
 
 #include "metadata.h"
-#include "metadata_hash.h"
+#include "metadata_segment_id.h"
 #include "metadata_raw.h"
 #include "metadata_io.h"
 #include "metadata_raw_volatile.h"
