+++ src/metadata/metadata_segment.c
