+++ src/metadata/metadata_segment.h
