+++ src/metadata/metadata_superblock.c
@@ -6,7 +6,7 @@
 #include "metadata.h"
 #include "metadata_core.h"
 #include "metadata_internal.h"
-#include "metadata_segment_ops.h"
+#include "metadata_segment_id.h"
 #include "metadata_superblock.h"
 #include "../ocf_priv.h"
 #include "../utils/utils_io.h"
